# src/server/openapi.yaml
openapi: 3.0.3
info:
  title: وسِيطي API
  description: Real Estate CRM & Platform Integration API
  version: 1.0.0
  contact:
    name: Waseety API Support
    email: api@waseety.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.waseety.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        phone:
          type: string
        email:
          type: string
        type:
          type: string
          enum: [lead, client, owner, tenant]
        status:
          type: string
          enum: [new, contacted, qualified, closed_won, closed_lost]
        source:
          type: string
        lead_score:
          type: integer
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
security:
  - bearerAuth: []
paths:
...
  /crm/contacts:
    post:
      tags:
        - CRM
      summary: إنشاء عميل جديد
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - full_name
              properties:
                full_name:
                  type: string
                  example: "محمد العتيبي"
                phone:
                  type: string
                  example: "0501234567"
                email:
                  type: string
                  format: email
                type:
                  type: string
                  enum: [lead, client, owner, tenant]
                  default: lead
                source:
                  type: string
                  example: "website"
                source_meta:
                  type: object
                budget_max:
                  type: number
                requirements:
                  type: string
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Contact'

    get:
      tags:
        - CRM
      summary: قائمة العملاء
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, qualified, closed_won, closed_lost]
        - name: assigned_to
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Contact'
  /crm/contacts/{id}:
    get:
      tags:
        - CRM
      summary: تفاصيل العميل
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contact details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Contact'

    put:
      tags:
        - CRM
      summary: تحديث العميل
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                status:
                  type: string
                  enum: [new, contacted, qualified, closed_won, closed_lost]
      responses:
        '200':
          description: Contact updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Contact'
  /crm/contacts/{id}/activities:
    post:
      tags:
        - CRM
      summary: إضافة نشاط للعميل
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - summary
              properties:
                type:
                  type: string
                  enum: [call, email, whatsapp, meeting, note]
                summary:
                  type: string
                  example: "مكالمة متابعة"
                details:
                  type: string
      responses:
        '201':
          description: Activity created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
  /crm/tasks:
    post:
      tags:
        - CRM
      summary: إنشاء مهمة جديدة
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contact_id
                - title
              properties:
                contact_id:
                  type: string
                  format: uuid
                title:
                  type: string
                  example: "متابعة العميل"
                description:
                  type: string
                assigned_to:
                  type: string
                  format: uuid
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                due_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
...